Add support for iterating over regions and accounts in CLI
The goal of this change is to provide a simple way to execute a particular
command in multiple regions and accounts. The implemented mechanism is roughtly
an order of magnitude faster in execution time than iterating using a shell
loop.

For account iteration, the organizations:ListAccounts facility is leveraged
to provide a list of accounts to iterate over. The sts:AssumeRole service is
then used to obtain access to the identified accounts using role switching.

For region switching, a client is created for each enumerated region. This
facility uses a new command line switch rather than overloading the
--region switch to allow the account iteration to function correctly.