refactor config tests to be in CI and use `run_cmd` instead of `aws`
Moving to functional suite so that they run in GitHub CI. They can be
moved because no actual HTTP requests are made using the `configure`
command. These tests either silently failed or weren't run at all in a
release that allowed a bug to be introduced to customers using the
latest version of v1.