Use include_package_data instead of package_data
This allows us to specify the package data once in the Manifest.in
file. Specifcally, the include_package_data will include files
listed as part of the Manifest.in as part of the package if they
reside in the packages directory. This will also help us avoid
issues where some package files are included in the sdist and
not the wheel because the package_data differs from what is
declared in the Manifest.in