Enable dependabot for v2 branch
The plan is to enable dependabot updates for each of our
dependencies one at a time until all dependencies are accounted
for. At that time, we can move away from an allow list.

Note that even though this is for v2, the commit is on the develop
branch because it is the default GitHub branch and GitHub only
pulls configuration from the default branch.

In the future, when we make v2 the default branch, we should be
able to just remove the target-branch setting.