Remove symlink resolution in Loader.is_builtin_path
Because the Loader class never sees the filesystem path of the data
file, only the data path (which excludes the file extension), the call
to os.path.realpath() in #2751 would have never resolved a
symlink. Even if working correctly, this symlink resolution would only
have taken effect for users who have a symlink in their data
directory (~/.aws/models) pointing back to the botocore package's
builtin file. After considering the benefit of covering this unlikely
edge case and the effort involved in making the filesystem path
available to the Loader class, the decision is to remove symlink
resolution from is_builtin_path.