Use source client for s3 --copy-props subscribers
When retrieving metadata and tags, the subscribers were using
the intended transfer client instead of the source client. This
lead to failures when calling GetObjectTags and unnecessary
additional region redirect requests for HeadObject when using an
opt-in region (e.g. af-south-1) that did not match the region
of the source bucket.