Fix parsing of dotnet output for `aws codeartifact login` command.
The login command for dotnet is unable to parse the output from the command
`dotnet nuget list source --format detailed`, resulting in an index out of
range error. This is because with dotnet tool version 6.0.402, a warning
message was added in the end of the output.

Additionally, if `DOTNET_NOLOGO=true` is set, a header message at the start of
`welcome to dotnet 1.0!` is added, which breaks the parsing.

Finally, responses from the `dotnet nuget list` command do not always come in
utf-8 format.

This PR addresses these issues to make parsing independent of any lines other
than those that appear in the Registered Sources section from the command
output.