Remove unused conditional to invoke subcommand
This is already handled with the SubCommandArgParser so
it's no longer needed.

The BasicCommand could also have the last elif clause
removed, but it's there as an extra safety net as
we'd by relying on the SubCommandArgParser to verify
that the subcommand table is not None instead of explicitly
checking this in BasicCommand directly.