Ignore errors setting always-auth in npm config set
The release of NPM 9 has dropped the support of deprecated or unsupported
parameters in the `npm config set` command. This includes the `always-auth`
parameter. Since `always-auth=true` is being set internally via the
`aws codeartifact login --tool npm` command, this results in users of NPM 9
getting an exception. this blocks users to use login command
entirely with NPM 9. This flag is ignored by NPM 7 and 8. Since the NPM login
command is also used for yarn users, it was not previously removed.

This change is to catch and ignore an exception while setting `always-auth`
flag for the `npm config set` command. Users with NPM < 9 will still set
the parameter in their `.npmrc` file, while users with NPM >= 9 will not.