Add pagination backcompat for kinesis list-streams
The ListStreams previously used the ExclusiveStartStreamName parameter
for input tokens to pagination. This operation was then updated to
also allow for the typical NextToken input and output parameters. The
pagination model was also updated to use the NextToken field instead of
the ExclusiveStartStreamName field for input tokens. However, the
ExclusiveStartStreamName is still a valid parameter to control pagination
of this operation and is incompatible with the NextToken parameter.

So, the CLI needs to continue to treat the ExclusiveStartStreamName as if it
is a raw input token parameter to the API by disabling auto-pagination if
provided. Otherwise, if it was treated as a normal API parameter, errors
would be thrown when paginating across multiple pages since the parameter
is incompatible with the NextToken parameter.