Force timezone info to fix possible test failure
This commit will affects the test
`test_credentials.py::SSOSessionTest::test_token_chosen_from_provider`.
This test will throw a `RuntimeError: Credentials were refreshed, but
the refreshed credentials are still expired` because the timestamp call
uses system local time, which is different from UTC time. And this will
cause an unexpected failure if the test is not run in a system configured
with `TZ=UTC`.

Signed-off-by: Avimitin <dev@avimit.in>