Fix incorrect --help output
This issue began in version 2.1.0. See
https://github.com/aws/aws-cli/compare/2.0.63...2.1.0#diff-45cb40c448cb3c90162a08a8d5c86559afb843a7678339500c3fb15933b5dcceR89.

A parsing layer (FirstPassGlobalArgParser) was added specifically to parse
global arguments. It is a subclass of argparse and by default it injects a
_HelpAction unless `add_help=False` is provided.

It also suppresses its help output, so the the --help action consumed the --help
argument. This led to the default arparse output and exiting with rc 0.

The global parsing logic was also moved into `create_clidriver`. This means that
all test logic using clidriver.main is no longer running arguments through the
full argument parsing gauntlet that is being run in real usage. So the tests
designed to target --help, --profile, and --debug are not going through the code
path that actually parses them.

Further review is needed to see which tests need to change and how the parsing
behavior should be updated.

modified:   awscli/argparser.py #	modified:   tests/unit/test_clidriver.py #