Add github actions to handle regenerating lockfiles
This adds a new job to the test source distributions workflow which
attempts to regenerate the lock files based on the requirements
files. If any changes are introduced they are commited and pushed to
the branch. Next the tests are run conditionally, if there were no
changes commited then the tests are run, if there were changes the
tests are skipped.

There is also a new workflow for generating a changelog from a
dependabot commit.