Consistently remove ResponseMetadata field in output
The CLI contains logic that removes the ResponseMetadata field
that is always populated in a botocore client response dictionary.
This ResponseMetadata field is never supposed to be present in the
output from the CLI command as it clutters command output with
metadata information not particularly useful to the general command
invocation, all which can be located in debug logs.

However, there was a small edge case where the output
would leak the ResponseMetadata field if an Errors field was
present in a sucessful API response. This behavior as it is today
is unintentional and undocumented. Originally, the logic was intended
to target a pre-GA interfaces of botocore that did not raise client
exceptions for non-2xx responses and therefore has not been
applicable to CLI output formatting since 2015.