Introduce auto mode for preferred_transfer_client
This is the new default value for the s3 ``preferred_transfer_client``
configuration setting. In ``auto`` mode, the AWS CLI will determine
whether the current system is optimized for the AWS CRT client
(e.g. is an EC2 environment with high network bandwidth).
If it is optimized and there are no other AWS CLI processes
using the AWS CRT client, the AWS CLI will use the AWS CRT
client, and if not, it will use the normal Python transfer client.

In introducing the ``auto`` value, I also renamed the preexisting
``default`` value that mapped to the normal Python transfer client
to ``classic``. Specifically, it was confusing that ``default`` was
no longer the default value for the ``preferred_transfer_client``
setting. The ``default`` value is now an undocumented alias for
``classic`` that avoids runtime errors for current users that
may have explicitly configured the AWS CLI to use the normal Python
transfer client.