Extend docutils line length limit
This is a temporary fix to handle JSON snippets in the man pages
that are longer than the default docutils line limit of 10,000
and cause the man pages for some EC2 examples to not render.

Long term, we should consider breaking up the enums to multiple
lines or potentially not enumerate all enums in the JSON snippets.

The value 50,000 was chosen based on the line length for
InstanceType enums for the ec2 create-launch-template-version
command. Currently, the line length for the InstanceType enums
is just over 10,000. Assuming each enum is 10 characters long,
a line length of 50,000 would give us the runway for roughly
4000 more instance types. The line length was not made magnitudes
higher (i.e. to avoid hitting this issue ever again) because
lines of this length are already not ideal and we should make sure
to implement a more long term solution.