Allow configuring local IP for open-tunnel.
The IP used for ec2-instance-connect open-tunnel was hard-coded to
localhost. This means that it does not work for users of the AWS CLI
Docker image, because Docker can only expose services that listen on
0.0.0.0.

This PR adds an option --local-ip, consistent with --local-port, which
allows the user to override the IP that the tunnel is bound to.