Fix pytest-8.0.0 test failures due to incorrect `pytest.warns()`
This is a port of https://github.com/boto/botocore/pull/3109

Remove two incorrect `pytest.warns()` invocations
for `.get_component('internal')`, in order to fix test failures
with pytest 8.0.0.  This version of pytest fixes `pytest.warns()` to
work inside `pytest.raises()`, and therefore causes the tests to fail
because the warning is not emitted when `get_component()` raises
`ValueError`.