Pin colorama in bundled installer
Updating to colorama > 0.4.5 requires additional
build dependencies (`hatchling`, `flit_core`,
etc). This increases the complexity of the bundled
installer.

We cannot add this to `EXTRA_RUNTIME_DEPS`
because the `pip download` in `download_cli_deps`
will fetch the latest `colorama`, which will take
precedence when running the `install` script.

This change pins `colorama` to the last version
without the additional dependencies.