Update RSA private key size to 2048 in test.
This is required to unblock updating our
`cryptography` pin:

https://github.com/aws/aws-cli/pull/8895

This test was ported 5 years ago as is using a key
size of 512:

https://github.com/aws/aws-cli/pull/4846

As of `cryptography==43.0.0` the minimum key size
allowed is 512:

https://github.com/pyca/cryptography/commit/83dcbc190165ad5c1f86bddaee76e0b288803c43

This change bumps it up to a more reasonable value
of 2048. This has only a couple hundredths of a
second time difference to generate as tested on my
Mac.