Add support for Role Session Name to EKS commands
This commit adds a --session-name option to the eks get-token and eks
update-kubeconfig commands. Prior to this, when creating a kubeconfig
file with eks update-kubeconfig and specifying a --role-arn, there was
no way to set a custom Role Session Name. This session name was
hardcoded to "EKSGetTokenAuth". Role session names are included in the
output of commands like `kubectl auth whoami` and can be used to
distinguish users connecting to the EKS cluster that are using the same
--role-arn.