Use CRT for cloudtrail verify query results
This commit makes two changes that were difficult to make
separately due to a change in the interface of signing and verifying
with public and private keys.

First, it migrates all remaining use of cryptography to CRT for the
CloudTrail verify query results.

Second, the existing test cases relied on utilities from the
cryptography package to create a temporary public key from a given
private key file. This has been refactored to move the public key to a
DER-encoded PKCS1 file that is similarly read in and used directly
through the utility module  `PublicPrivateKeyLoader`. A DER-encoded file
was chosen because it is the format used in the API, and it could easily
be read in directly without other cryptographic file manipulation. The
`PublicPrivateKeyLoader` class is now used in the CloudTrail log
validation tests to keep them as similar as possible.

I left the data files separately for each test suite (unit and
functional), even though they are identical since there was no
preexisting structure for sharing data files across the test suites.
This entailed duplicating the helper file loading functions as well,
since they relied on the test module directory at import time.