Add pre-commit hook
First pass at adding a pre-commit hook to enforce
formatting and linting standards.

This adds the pre-commit command to run on a
limited set of directories. Subsequent changes
will remove excluded folders and run formatting on
them as well.

The command will use the standard pre-commit hooks
for yaml, whitespace, and end of line linting.

It also uses `ruff` configured identically to
the develop branch, except for disabling
unused import checks ("F401"), as there are known
imports that are unused in the module but imported
in other modules. This is common in `compat` but
in other places as well.

It also uses `isort`, also configured identically
to the develop branch.

The intention is to start enforcing formatting and
linting standards on edited code. Running the
pre-commit hook manually will make numerous
changes that should be broken into multiple
phases. Additional PRs will work to get the
codebase up-to-date passing all standards.