Add a console_scripts entry point for aws
In comparison to `scripts`, setuptools's `console_scripts` mechanism is
a more portable way to register CLI executables which allows setuptools
to generate the executable stub script in a manner appropriate for the
platform in question.  It also integrates well with bazel.  In this
patch we add one for the `aws` command.

Fixes #5063