Add configuration to disable host prefix injection
This change deviates from the upstream `botocore` to add the ability
to disable jost prefix injection. It changes the default for
`inject_host_prefix` in the Config object to be `None` (from `True`) so
that the value can be determined using the configuration precedence. The
fallback if the value remains `True` if it is set to `None` to preserve
the CLI default behavior.