Assume Role to find EKS cluster for config update
This commit will assume a given role when trying to update the
kubeconfig.

While `aws eks update-kubeconfig --role-arn...` would add the required
parameters to the user section of the configfile, but will fail finding
the cluster if the original profile does not have access.
This commit will fix the issue and assume the role before getting the
cluster description.

Fixes #8554