Remove unused lines.
In `tests/functional/botocore/test_endpoint_rulesets.py`, the lines
removed are used upstream to mark tests to be skipped. This is not
required for AWS CLI v2 since AWS CRT is always installed as a
dependency:

https://github.com/boto/botocore/blob/a867a74b0d003c8cb8bfb4abc8f6e6d35bb792e4/tests/functional/test_endpoint_rulesets.py#L174-L189

In `tests/unit/botocore/test_credentials.py`, these lines are simply
unused, and are not upstream either. Safe to remove.