Update sts endpoint assertions
The default behavior for which STS endpoint to use is changing from global to regional.

These tests assume that the global endpoint will always be used, which is not the case
for the default behavior from the botocore client. This change updates the tests to reflect
that the regional endpoint will be used.