[v2] Perform copy with no overwrite using IfNoneMatch (#9910)
This commit rolls back changes made to force multipart copies if no
overwrite is specified. This includes checking for 0-length files.
It  also removes the CopyObject argument block list, since IfNoneMatch
is now allowed and was the only argument in the list.

It also removes the tests for 0 length files and moves some tests to the
TestNonMultipartCopy class, since CopyObject accepts IfNoneMatch and no
longer needs to force multipart copying.